Mostly Implemented Square Subdivide.
The Triangle subdivide is fully working, but that is just the
same as the article. The square subdivide is not fully working,
as the original corners of the cube are causing issues. I think
this is coming from the difference in normalization in the
GanerateMeshSquare() method from the GenerateMeshTriangle() metod.

Hopefully the difference in OS this was build on does not cause
any issues. I guess we will find out. All further development
should be done in Windows 10 to keep this consistant.