Subdivide Cube Implemented.
Alright, the cube is fixed. Issue was that we were not normalizing
the Vector3s when creating the original cube mesh. With that working,
we have started to make a Planet class for PlanetFactory to output to.
The current Planet generated is blue with a red tile showing the path
of Polygons in sequence. That sequence is useless.

Moving forward, we need to get those Polygons into a usable pattern so
that we can actually apply Perlin noise to it later. For now, I am going
to look into how to map coordinates to a sphere and think on ways to get
the Polygons mapped to those coordinates. No idea on that yet.