Planet Classes cleaned up.
The Planet classes have been seperated into three classes. The
Factory makes the planet, Planet is is planet, and Controller
tells the facotry when to make them. Seems more flexible this
way, and is far easlier to maintain; or expand in this case.

I have included a noise generator that I will hopefully be using
to give these planets some character. Nothing has been done with
it yet, but the documentation is out there.