Some planet action happening.
The noise map is set up and in use, that's pretty neat. Due to
how we generate the planet, there are currently only 999,999
unique planets. Quite a few, but not quite enough.

Moving forward should be either continuing with making the
planets a bit more interesting, or making the tiles selectable.