Some Raycasting Added.
The planets have MeshColliders, and a raycast is able to hit it now. I
still need to add a mouse click as it current tracks on mouse hover.

The planet doesn't do anything with the raycast, and I am not able to do
much with the Polygon it hits yet. I think this is progress.