Updating polygon color based via raycast.
This is actually a pretty big jump. Using the raycast, we can turn entire
cubic polygons white. This should be done with a click as a selector, and
we still need a way to store the info on the poygons isntead of just over-
writing the color.

I would think the next step is to create an array that shares the triangle
indexes to store said info. Maybe I should clean up the current code to
prepare for that.