Display area working!
Alright Boshart, I take back the "curse you". The barycentric
part that was added in the last push worked perfectly. It is
lightening fast and flawless, well done. You taught me well
enough that I can still do it.

The problem was that I was supplying it the wrong values. The
triangle.HitIndex uses the triangle index from the mesh, and I
was supplying it the index from our local primitive. We also did
the vertices incorrectly when generating the original mesh, the
triangles were dependent on each other so the array was filled
with 30% garbage.

Working now though, and working great. Next step will be to make
the mesh on the planet fit the noise better, I look forward to
seeing what I come up with there; because I have no ideas yet.