Improved Planet Generation.
The planet texture now more accutaly represents the terrain, I
thought that would be more difficult. The seed also generates
a random axis and sort of generates a subdivide on the planet.

I want subdivide to have more of an impact than it does, but it
has a bigger issue at the moment. I can't subdivide a planet
with a value of 6 or higher. I need to look into that.